+++ CHANGELOG.md
@@ -30,6 +30,8 @@ be enable with an 2nd Optional Boolean Argument as `true`
 - added a BUILDS file and changed some stuff in the README.md
 - added Workflow to zip the documentation and upload it to the next Github Workflow
 - fixed some paths in the IO and the OS Library
+- fixed Workflow for Windows Builds
+- fixed ZIP functions and added Tests to them
 
 Read the DOCS for Version 0.2.0 for more Infos